逐步替换LiveData为SharedFlow

+++ repository/src/main/java/com/senierr/repository/service/impl/CommonService.kt
@@ -1,6 +1,7 @@
 package com.senierr.repository.service.impl
 
 import com.senierr.base.support.utils.CloseUtil
+import com.senierr.base.support.utils.EncryptUtil
 import com.senierr.repository.disk.DiskManager
 import com.senierr.repository.remote.RemoteManager
 import com.senierr.repository.remote.api.CommonApi
@@ -81,11 +82,11 @@ class CommonService : ICommonService {
                     // 判断文件是否存在
                     if (destFile.exists()) {
                         // 判断是否需要重新下载
-//                        val destMD5 = EncryptUtil.encryptMD5File2String(destFile)
-//                        if (destMD5 == md5) {
-//                            continuation.resume(destFile)
-//                            return@suspendCancellableCoroutine
-//                        }
+                        val destMD5 = EncryptUtil.encryptMD5File2String(destFile)
+                        if (destMD5 == md5) {
+                            continuation.resume(destFile)
+                            return@suspendCancellableCoroutine
+                        }
                         val result = destFile.delete()
                         if (!result) {
                             throw Exception(destFile.path + " delete failed!")
